voiceops/voiceops-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic
prometheus-client
python-multipart
openai
python-dotenv